app/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Alexander van Zyl <vanzyl.alex@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
fastapi = {extras = ["all"], version = "^0.54.1"}
inflect = "^4.1.0"
sqlalchemy = "^1.3.16"
psycopg2-binary = "^2.8.5"
sqlalchemy-utils = "^0.36.5"
argon2-cffi = "^19.2.0"
python-jose = {extras = ["cryptography"], version = "^3.1.0"}

[tool.poetry.dev-dependencies]
pytest = "^5.2"
flake8 = "^3.7.9"
isort = "^4.3.21"
sqlalchemy-stubs = "^0.3"
autoflake = "^1.3.1"
autopep8 = "^1.5.2"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
